Add a small parser to convert from/to Json, and adapt Fake to it
Be able to parse from JSon all the types we need, we can't simply use
QJson because it only knows about QObject and some of our properties
aren't such QSize or QPoint or QRect.

With this, we have also added proper parsing of modes and sizes (no
longuer a dummy object).