Remove the execution of Xephyr in the test, we can't do this this way
When QApplication is executed the DISPLAY env var is still :0 so it
connects to the :0 X11 server. There is no way to change that connection
without restarting the application so we won't be able to make this test
pass on Xephyr doing it this way.

So what we are going to do is somehow from cmake output a bash script
that will prepare the environemnt (use DISPLAY = :1 and execute Xephyr).