Pass the test using actual XRandR :)
I'm not sure if the implementation is complete but it passes the test
and it tests all that can be tested with Xephyr (single output, and all
properties).

Writting this backend I realize how stupid this architecture is, we are
basically copying everythign in an endless group of FOREACH wasting
time and being stupid.

I don't want to invest more time on architecture right now, the external
api is good enough (Config, Output, Mode objects) to use.

I will refactor the architecture completly probably using interfaces once
the KDED and the KCM are done.