Unhardcode the crtc used in changeOutput and enableOutput
For changeOutput we are using a cached crtc, since if we try to get it
using outputCrtc it will (or may) return 0, since we may have called
disableOutput before calling changeOutput.

For enableOutput we are using the just added freeCrtc, which checks for
a not used crtc and returns it.

For some reason I can't understand if we disable an output using the
crtc 64 and try to enable it agin using the crtc 63, it will fail. So
we are force to reuse the previously used crtc (that's why we are
caching them in currentCrtc).

I will try to figure out why this happen and test if I can enable 2
outputs with the current code