Set parent to config objects.
ATM I'm setting the object Config as a parent, I still don't know if
we want to have a Config that is a tree, or  not.

In case of a tree it would be:

Config
  |
  +--Screen
       |
       +--Output#1
       |    |
       |    +--Mode#1
       |    +--Mode#2
       |
       +--Output#2

That tree would be the more technically right approach but I'm not
certain that it will turn out in a good api, or an api that will use
most of the people, it is easier to do:

Confing->Outputs

Rather than:
Config->Screen->Outputs.

Since we are NOT suppoting multiple screens (ATM) we don't really need
the tree imho.

Anyway, we will see.