Fix handling of disconnected outputs in updateOutput()
When an output is disconnected, we receive RRScreenChageNotify and
RRNotify_OutputChanged events, both causing the config to be updated.
There was an issue with currentMode being still set in RRScreenChangeNotify,
however isConnected being false, thus we had currentMode pointing to an
invalid mode. This is no fixed by not updating currentMode of disconnected
output.

Together with previous commit, this should eliminate unnecessary property
change notifications from KScreen::Output.