Update screen size when we fail to enable an output
We call XRRSetScreenSize() before we try to enable outputs. However
when for some reason XRRSetCrtcConfig() fails and output is not enabled,
the screen size is still set to the previous value.

In case we failed to enable the {left,top,right,bottom}-most output,
we screen size we set is bigger then the actual screen size, so in such
case we update it to match the current settings.