Use our own X Window to select input so we don't conflic with others
XRRSelectInput select's input per X connection and Window, so if we want
to be a KDED module we have to either create our own connection to X
(really ugly) or our own Window (still ugly but Xorg is like it is).

So, with this commit we use our own Window (which ofc is not shown to the
user in any way) so nobody can smash our Mask or we can smash other's.