Fallback to the biggestMode when no preferredMode is given
Brokens screens that provide no EDID will fail to indicate their
preferredModeId.

Since this is a problem common on all platforms (it does not matter if
we use wayland, xorg, quartz or framebuffer you will still have screens
with no EDID) the right place to solve this is in the abstraction.