First import of XRandR1.1 backend using Xcb
Looking at the best approach to support XRandR 1.1 I realized that the
best solution will be having another backend that takes care of it.

XRandR1.1 is much much simpler than XRandR > 1.2 and if we want to
support it in the current xrandr it will mean basically doing a new
xrandrconfig specifically for it, so it is clear to me that the best
thing to do is just creating a new backend.

This backend will use xcb instead of xlib and I hope it will be my
training on writing a new xrandr (>1.2) backend using xcb.

Oh, and we are using the nice wrapper done by the KWin project that
makes xcb suck less.