Use the information from currentOutput, not from the unapplied config
Before this patch, we were looking for the size of the current Mode
in the Output object of the config that is passed to Config::setConfig
which either can or can't have the wanted Mode.

This has been kinda working so far because and ONLY because in
Generator (from kscreen/kded) we use the currentConfig as the base for
the config that we are generating, so all modes are included.

CCMAIL: dvratil@redhat.com

BUG: 320537
BUG: 321708
BUG: 321699
FIXED-IN: 1.0.1