Update XRandROutput after it's disabled so that cache is kept up-to-date
After disabling an output, we have to force-update the XRandROutput, because right
after that we receive RRNotify_CRTCChange signal for that very output. This is
propagated to users (like KCM), which then call canBeApplied(). canBeApplied()
compares the KScreen::Config with the *outdated* XRandRConfig and that leads to
a crash (because the XRandRConfig still claims that the output is enabled).

By updating the cache right after we disable the output, when KCM calls canBeApplied(),
it's Config is compared to an up-to-date Config and everything is fine.

REVIEW: 117298