Add a small dirty copy of KScreen console to test the lib
KScreen is not ported to frameworks so I can't test the lib using
the console. So I have copy pasted the most important part of it
so we cand o testing of the lib only compiling the lib.

Further improvement will be removing the code duplication by
implemnenting qDebug methods for our classes.