Block until all objects are initialized
In WLConfig's ctor, we're initializing the connection, connect to the
Registry::sync() signal. In the initialization phase, we're blocking the
ctor with an event loop, which is terminated when we're done enlisting
outputs and all outputs have been fully initialized (Output::complete()
signal).

This involves a bit of book-keeping, we need to keep track of what's
done initializing, in order to make sure that everything is complete
before we're return from WaylandConfig's ctor, and the library user
assumes that everything is loaded -- at this point, we can produce a
Config* object that is in sync with the current state.

This is needed to produce a useful Config object for our clients. It
means that we have to block in the Config ctor while listening to the
signals announcing outputs, and for the outputs to complete
initialization (their modes, geometries, etc.).

Needs kwayland[sebas/wl_display_sync] to build.

This patch needs some improvements, error pathes aren't correctly
handled for ouputs, for example, but it allows the current tests to all
pass, so it's a good start.