Create and start our own Wayland Server
We start a test-internal wayland server, so the wayland libkscreen
backend can connect to that and run the tests. This ensures a full
roundtrip, and can be nicely extended.

Starting and stopping the server is done once for all tests.

This needs the KWayland/Server headers installed, which is not currently
the case. You can use the kwayland[sebas/waylandserver] branch for this.