Introduce new asynchronous API to get and set KScreen Config
The old Config::current() and Config::setConfig() methods were removed,
instead GetConfigOperation and SetConfigOperation classes are introduced
here. The emit finished() once the operation has completed. The operations
will auto-start and auto-delete.

The reason for having the async API is that BackendManager might need to start
the BackendLauncher first, and then still a DBus call is involved to get data
from the backend (which might also block for a moment). This would never be
acceptable to be called in blocking manner from main thread, hence nice simple
async API has been introduced.