BackendManager: reset crash counter after a minute
The point of the crash counter is to prevent application from trying to
restart the backend in an endless loop in case the backend crashes every time
after start. However occasional crash is ok (look away kids, you'll understand
this when you grow up), and in case of a long-running session this would make
the KDED module unusable, as it would refuse to start the backend after just
4 crashes.

This change implements a simple timer, that will reset the crash counter to 0
if the backend manages to run for at least 60 seconds without a crash.