ConfigMonitor: don't request config when initial backend is received
KScreen KDED requests a config immediatelly after start, then compares it to
it's config files, requests change if necessary and then listens for change
notifications to write the new state down. However the ConfigMonitor is only
first initialized after the config is received by KDED, so onBackendReady is
called around the time KScreen has already requested the changes to be made.

If we would request a config from onBackendReady now, then a race condition
would occur later, when we get one config from the change notification and
the config requested from onBackendReady, which would however be out-of-date
at this point (due to how backends work). This means that Monitor would first
update the configs to the new state, and then subsequently update it back to
the original state.

To avoid the situation above, we don't request a config from onBackendReady
when the backend we received was the "initial" one, since there is no reason
to do so. The request makes only sense when we already had a valid backend
and received a new one, which means that backend probably crashed and we
might need to get in sync with the reality again now.