Fix crash when multiple EDID requests for the same output are enqueued
When the first EDID date are delivered, the output is removed from
the list of pending requests, causing crash when response to the second
request arrives.

This fix makes the code more robust by allowing multiple request for the
same output per single config as well as handling multiple configs.

BUG: 343482
FIXED-IN: 5.2.1