Factor out config reading
Transplant the wl output code to the wayland backend dir to put it the
reading and the writing code right next to each other as to reduce the
tendency to not update the reader when changing the writer.

This is built separately and only from the tests. This way we avoid
introducing a KWayland::Server dependency into libkscreen -- it only
needs the client.