[wayland] Add and hookup configwriter
With this patch, the wayland backend writes out an incomplete config
file. The outputs are pretty empty right now, but the resulting json
should be compatible with or have at least a subset of with the
configurations under autotests/configs.

Implementations and autotests to follow, but the basics have been laid.