Add Output::size and make it report correct size
We were making some assumptions in Output::geometry like for example
assuimg that the resolution of the currentMode was actually the output
size which is not always the case, like for example in a case we were
already covering of vertical transformed output.

In order to fix this proplerly we are adding a size property to Output
that backends and fill with whatever value they want and we use it to
create the QRrect with return in Output::geometry.

QScreen: Whatever QScreen::screenGeometry().size() returns
Fake: Added a test for singleOutput, the rest is all the same
XRandR11: Whatever size has currentMode

XRandR: Whatever size CRTC reports.
Note that crtc is the one in charge of applying transformations so it
always reports correct geometry, even on scaled or vertical/landscape
outputs.