Fix regression in Output::geometry() behaviour introduced in f57dc
Output::geometry() always reflects whatever is currently stored in
the Output (which can be different from what is the currently active
configuration). This was broken in the previous commit, because the
size() property is only updated when we get update from KScreen backend.

The API documentation for those properties has been updated to
explictly and clearly describe the difference in behaviour of those
properties.

CCMAIL: afiestas@kde.org