XRandR: use intermediate screen size when applying config
We apply the screen size before we do any changes to outputs. In some cases the
screen size for the new config is smaller than the current screen size (like when
rotating an output), and it can happen that XSetScreenSize fails or applies only
partially. To workaround that we first set an intermediate screen size which can
fit both the current and the new config, then apply the output changes and then
(if necessary) update the screen size once more to fix the new config precisely.

This should fix some weird issues with screen size when rotating an external
output.