Use DBus activation to start the backend launcher
The custom QProcess handling and stdout-based communication was very fragile
and caused various issues e.g. when a Qt plugin was writing to stdout as well, it
confused BackendManager and blocked entire plasmashell start.

This patch changes the BackendManager and BackendLauncher to use DBus activation.
The backend launcher always registers org.kde.KScreen service (which means that
we can now only run one launcher, but that only really affects KScreen developers)
 and has org.kde.KScreen interface through which BackendManager can request the
launcher to load a backend. The launcher loads a backend and registers its
/backend/org.kde.kscreen.backend intererface.

As a side effect of this change, config retrieval on start seems to be much
faster, because we don't have to execute the kscreen_backend_launcher just to
find out there's alread another instance running and what's its DBus address.

Tests are now executed in isolated DBus environment via dbus-launch to ensure
that they don't clash with each other (in case of parallel execution) or don't
clash with the main launcher running on developer's machine.

BUG: 353720
BUG: 353685
FIXED-IN: 5.5.0