Load backend in InProcessConfigOperation
This is the gist of loading the backend in-process. Rather than going
through BackendLoader, this new ConfigOperation loads the ConfigPtr
directly through QPluginLoader, rather than dbus-autolaunching the
backend out of process.

Why?
- backends in-process are easier to debug. It's already a vast
  improvement getting
- additional isolation of backend types -- this solves the case where
  the backendloader is already running and we're trying to load a
  different backend.
- performance, not starting a separate process and marshalling
  everything on both sides cuts out a lot of IPC overhead.

This change is inspired by the Wayland port, but not exclusive to it. We
have tight control over the tools, and also the protocol level bits are
much leaner, so in that case, the overhead of the out-of process
approach can be avoided, with some additional benefits.

Edid and writing still need fixing.