clear pointer before init
This allows to call init more than once, which is useful for
reinitializing the backend in case it's used with multiple setups, such
as in the autotests.

Now, on subsequent calls into init, the arguments may have changed, so
we're re-creating the whole config from a new json file.

Makes a whole bunch of tests pass with KSCREEN_BACKEND_INPROCESS=1 set.