Support options for both kinds of jobs
This change moves the Options enum into the ConfigOperation class,
making a cleaner separation of interface and implementation.

This also allows us to pass options, such as NoEdid down to the
backends.

I'm not sure if this is all binary-compatible, it is source-compatible,
and the new mechanism is only used opt-in, by using the new
ConfigOp::create(..) API.