in-process mode merge
Load backends in-process when KSCREEN_BACKEND_INPROCESS=1 is set in the
environment. Defaults to out-of-process mode.

purpose:
- allow easier debugging
- for some backends (qscreen, upcoming kwayland) the out of process
  operation is not necessary since these backends are well-shielded

This implementation is backwards compatible and opt-in for running
setups.

If the user exports KSCREEN_BACKEND_INPROCESS=1 before running, all
operations will be done in process. Otherwise, the out-of-process mode
is used.

Changes revolve around the ConfigOperations, the BackendManager (which
"factories" the backends, either in- or out-of-process), and the plugin
loading logic

Autotests should cover all the cases (and actually a few currently
unsupported ones, such as using different backends in the same process).

Details on performance, etc.:
http://vizzzion.org/blog/2015/11/wayland-and-libkscreen-benchmarks/
This is also the merge commit for the in-process feature.

REVIEW:126101