XRandR: handle changeOutput() without CRTC better
It is possible that XRandROutput counterpart to KScreen::Output does not have
a CRTC set while the KScreen::Output reports itself as enabled. This leads to
changeOutput() hitting an assert, because we assumed that such case can never
happen. Now we know it can, so instead when changeOutput() is called for an
output without CRTC, we fallback to calling enableOutput().

BUG: 344813
FIXED-IN: 5.5.5