We need to decide early on which method to use
The ctor sets some stuff based on the method, if we call setMethod()
afterwards, it has to go through initMethod() again, but it's still too
late, so we run into the case where OutOfProcess is initialized but
subsequently, InProcess is used.

I'm not overly happy here, as

- we're making the decision which backend to use really complex (the
  continue logic doesn't help here)
- we're starting to duplicate this logic (not in the ctor)

This prevents a crash though, so let's run with this for now, and build
up courage to rewrite this part.