backendloader defaults to in-process loading unless x11
This change prefers the in-process mode to be used for all backends,
except XRandr. The KSCREEN_BACKEND_INPROCESS env var still bears highest
priority, but if it's not specified, we'll now use the in-process mode
by default for wayland, qscreen and fake backends, since these are
considered safe (enough) for in-process operation.