Set output name to model name
This aligns our output names with QScreen's names. Plasmashell assumes
that libkscreen's Output::name() is the same as QScreen::name() for a
given output. With the QScreen backend, this is of course always the
case, the xrandr backends expose the same behavior, so align kwayland's
behavior with that as well.

Long story short: This fixes plasmashell startup in a wayland session.

Unfortunately, we can't really auto-test this yet, since our test server
isn't able to host an app with QT_QPA_PLATFORM=wayland, which is needed
to load QGuiApplications::screens() to compare the screen names with
QScreen's. I've tried.