Allow preferring a specific backend in backendloader
Summary:
The name argument to loadBackendPlugin is used from the external backend
launcher, but was not considered when loading the backend plugin. That
means that loading the Fake backend using the out-of-process mechanism
failed, since the backend isn't passed and the runtime-detected plugin
would be loaded - that's never the Fake backend.

This patch passes the name argument into preferredBackend, and if
supplied, it's the preferred filter for the installed backend list.

Test Plan:
* added autotest
* ran kscreen_backend_launcher Fake" on the command line, verified that
  the Fake backend is loaded

Reviewers: graesslin

Reviewed By: graesslin

Subscribers: broulik

Differential Revision: https://phabricator.kde.org/D1325