be lenient about physical size in qscreen backend tests
The xvfb server on the CI system doesn't report physical size, so avoid
a complete test failure, since it's the X server that's broken, not the
test.

I've tried getting the qscreen backend test to run against our wayland
testserver -- and thus through the wayland QPA -- but that introduced
large pieces of boilerplate in the server, and more complexity in the
autotests, when we create the test server from a QApplication it races
since we can't create the wayland test server before the QApplication is
created (it pretty much needs a QWidget), and the QApplication init
fails when the wayland server is not fully started. Seems just not worth
it...