validate the config before checking if it can be applied
Summary:
canBeApplied is public API, it shouldn't blow up when passed an invalid
config, but indicate (correctly!) that this "config" can't be applied.

This is called a bit carelessly from the kded daemon. I've fixed it
there as well, but it seems prudent to also make sure in libkscreen that
the user doesn't do stupid things.

CCBUG:362586

Test Plan: autotests pass

Reviewers: #plasma, graesslin

Reviewed By: #plasma, graesslin

Subscribers: graesslin, plasma-devel

Projects: #plasma

Differential Revision: https://phabricator.kde.org/D1532