implement and document the new Log class
The general idea is that we cache file names and initialization from the
env vars on first creation, allow subsequent context changes.

The log file is opened and closed right away, since other processes may
want to also write to it.

We don't do any file-locking to keep things simple. The logging isn't
mission-critical, but rather a debugging tool for users. If we end up
with corrupt logs, that's bad luck (but can happen due to multi-process
race conditions). More advanced multi-process locking and delayed
writing is just going to give us headaches, so we avoid that at the
price of logs not being 100% reliably formatted.

Also, the code isn't robust yet, a couple of todos are inside.