xrandr-crtc: clear outputs in update()
In update, we initialize the entire chain of connected outputs. Since
update() is now called more than once during the life-time of the crtc
object, we need to clear it before filling the vector containing the
connected outputs.

This leads to the problem that crtcs are not freed up, which means we
may run out of crtcs quickly if a display is connected and reconnected.

This fixes a regression introduced with c2e1becccec.