[kscreen-doctor] support identifying outputs by name
This allows to use config commands such as "output.HDMI-1.enable"
instead of "output.73.enable", the former being more human-readable,
easier to identify, and more stable (the output id may change across
sessions).

If an output with that name isn't found, we still parse it as an int
output id.