fix mode setting when output is already enabled
This fixes setting mode when the output is enabled. In this case, we
have to pick the mode of the KScreen::Config passed in. This falls
back to preferred mode, which will practicably always be valid.

If we consider the xOutput first, we pick the currently set mode and use
that -- hence, nothing is going to be changed. If the display went
through the enabled code path first, the xOutput won't have a mode
already, which is corrected at that point.

This fixes a regression introduced with f53805676d.

Thanks Raymond Wooninck for reporting and testing.