add flag to for per-display scaling
Summary:
This adds a KScreen::Config::Feature for per-display scaling. As we plan
to introduce this with Wayland, we need to be able to check in the UI
whether this feature is supported.

This will be implemented by the backends at a later stage, but can now
already be used to work on UI changes. A relatively simple change that
I'd already like to get merged.

Test Plan: added autotests for the new flag and possible combinations

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2866