Fix Output::geometry() when dealing with a scaled monitor
Summary:
When a monitor is scaled the "size" is not the same as the mode size, we
want things in global co-ordinates.

This is only relevant on Wayland, but on everything else scale is 1, so
this makes no difference.

Test Plan:
Plugged in normal monitor to laptop, generator had the correct co-ordiantes
of second monitor

Reviewers: #plasma, graesslin

Reviewed By: #plasma, graesslin

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D5697