start of a small tool to enumerate and rotate touchscreens
This will eventually be turned into backend and frontend-split code in
KScreen, the goal being to rotate the touchscreen along with rotating
the screen itself, so coordinate still match the mapping, i.e. touching
the screen actually generates pointer etc. events where the display
sits.

There is a number of caveats here:
- input device rotation is display server specific, I'm working on a
rough X11 implementation here, for kwayland, we'll need something done
through KWin, but since rotation of outputs (and buffers) isn't
implemented in kwin yet, it makes sense to concentrate on the KScreen
API and X11 bits, first.
- mapping between touchscreens and outputs isn't clear and not automatic
at all. The plan is to first support the most common case (one output,
one touchscreen, touchscreen sits atop the only output), and then move
on to more complex cases with multiple outputs and even multiple
touchscreens. My idea is to add a TouchscreenPtrList property to
KScreen::Output which gives access to the known Touchscreens per Output
and, perhaps also for setting and storing this mapping. Otherwise, when
rotating an output, attached touchscreens are rotated along with the
output. This should allow us to do the rotation atomically, as really
the last thing we want to have it rotating the touchscreen before or
after the output is rotated and have all kinds of events (moves,
presses, releases, etc.) arrive at the wrong coordinates -- this could
easily lead to disaster.

So, small test app first, then expand this into API and backend-specific
code. The current state has some boilerplate, plus it lists touchscreen
devices.