XRandR: Clear EDID data when monitor is disconnected from an output
Summary:
This fixes KScreen behavior in situation when a monitor is unplugged
from an output and another monitor is plugged into the same output
afterwards. By keeping the old monitor's EDID around, the KDED module
would then mis-identify the newly connected monitor as the old one and
would apply the wrong config.

By clearing the EDID we force the backend to get an up-to-date EDID
from the new monitor, once ConfigMonitor requests it.

Reviewers: sebas

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D8840