Wayland: don't share the same Config between backend and clients
Summary:
Sharing a single config leads to undesired side-effects. The GetConfigOperation
is assumed to always return a new copy of the current config that is not changed
unless registered to ConfigMonitor. When sharing a single copy this assumption is
broken which leads to misbehaviour in KDED.

This change makes GetConfigOperation always return a new copy of the config that
caller can register to ConfigMonitor if they want to, otherwise it remains
unchanged until released.

Reviewers: sebas, davidedmundson

Reviewed By: sebas

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9904