Wayland: don't update output configuration if there are no changes
Summary:
This prevents the code from unnecesarilly emitting configChanged() if there
was in fact no change. This could lead to an endless loop when a client applies
current config and despite there not being no changes the backend emits
configChanged() to which the client reacts by applying the current config again
etc.

Reviewers: sebas, davidedmundson

Reviewed By: sebas

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9906