Wayland: add support for output type detection
Summary:
We can share the same "guessing" code with the XRandR backend here in order to
detect what type the output is (depending on having some of the well-known
connection types like HDMI, DP, etc. in its name).

The main benefit of this change is that we can know detect built-in laptop
screens on Wayland.

Reviewers: sebas, davidedmundson

Reviewed By: sebas

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9908