Make TestInProcess skip out-of-process tests if D-Bus service uninstalled
Summary:
With the movement to support running unittests pre-installation and KDE CI
having adapted to that for some build setups, this currently breaks tests
which rely on subject-under-test D-Bus services being auto-started from
D-Bus service files, given the D-Bus daemon does not see the uninstalled
ones.

One solution might be to catch that situation and do the start of the D-Bus
service ourselves, but I failed to get this done quickly. So as intermediate
solution, to at least have the in-process tests no longer being covered by
the out-of-process ones failing, this patch will just skip the
out-of-process tests if the D-Bus service could not be started.
That follows the current behaviour of e.g. the KGlobalShortcutTest from
KF5's kglobalaccel.

Test Plan:
TestInProcess no longer fails if the org.kde.KScreen D-Bus service cannot be
autostarted.

Reviewers: dfaure, sebas, gladhorn

Reviewed By: gladhorn

Subscribers: gladhorn, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D13871